Add files via upload
Executive Summary

Bottom Line Up Front: We have developed an algorithm that accurately predicts house prices in the Arizona housing market. We examined and analyzed over 12,000 sales records of properties in Arizona and then used the insights to determine the most important criteria for assessing the sale value of a property which we found to be the approximate square feet and the price per square feet. However, in a bid to develop a working method with the highest degree of accuracy, we also investigated other characteristics of the properties such as the days the property had spent on the market, the year it was built, the number of bedrooms and bathrooms as well as the list price. The combination of all these resulted in the algorithm predicting sales prices for houses within an average of $11 above or below the actual sales price.
Analysis: The process of developing our algorithm started with us collecting observations that contain details of houses listed for sale in Arizona as well as their respective agents/agencies, features, and other relevant information. We then analyzed each of these features and tried to establish if there was a relationship between them. In instances where we found a strong relationship between two or more features, we ended up using only one to avoid difficulty with the interpretation of algorithms results. In a bid to improve the prediction accuracy, we ensured that there was enough variation in the data we used and that the same data points were available for all the properties being evaluated.
Then we proceeded to try to establish a common pattern among the houses by looking at how we could group them together either based on their size, the year it was built, the county and city they are domicile in, and the amenities they come with such as pools, horses etc. We discovered that the value of the house is also largely affected by the city it is in as well as the amenities it has. To improve further on our algorithm, we developed new features from the existing ones we had available. This process led us to the discovery that the product of approximate square feet and the price per square feet of a property would significantly improve the predictive power of our algorithm. This singular feature we created significantly improved our property sales price predictions. The difference between our predicted prices and the actual selling prices was reduced from an average of $50,000 to about $11.
While we have full confidence in the algorithm to accurately predict the sale prices of Arizona properties, we advise caution when applying it to the general Arizona housing market. This is because we discovered that of the 15 counties in Arizona, only one – Maricopa, was well represented in the data and as such it might not be helpful in accurately determining the sales price of houses in the other counties.
Conclusion: Given that the average selling price of the houses used in our analysis was $897,743 and the average difference between the prices we predicted, and the actual price was $11, we are confident that our model will perform excellently in predicting the selling price of new houses. Real estate companies, property investors, and individuals can utilize this algorithm to accurately predict the sale prices of houses in Arizona and make informed decisions about buying and selling properties or assessing the value of current holdings.